requirements.txt
pyrogram>=2.0.106
motor>=3.4
python-dotenv>=1.0.1
tgcrypto>=1.2.5
flask>=3.0
